cmake: mingw implib conversion, fix C++ harden flags
Added a new helper function (see MingwCompat.cmake) to convert MinGW
import libraries to Visual Studio compatible implibs as a post-build
step. This is necessry for users to be able to link against mstdlib
compiled with MinGW from a project that's using Visual Studio.

Also, fixed Harden.cmake so that hardening compile flags are also
applied to C++ code (if in use in project).