formats/csv: add M_csv_parse_add_headers(), fix quote bug
Added function to allow the caller to prepend a header line before parsing the data. This is useful when parsing a single CSV table that's been broken up into groups of rows, where only the first group has a header line.

According to RFC 4180, an embedded quote char requires the value to be quoted. Updated output code to do this.

Also, updated unit tests to actually test CSV parsing and quoting rules, the tests weren't doing that before.